test_cp_inode_invalidate: Rename the helper file, remove debugging.
